start optimizing useGetUserMarketInfo
