+++ src/pages/Dashboard/Markets/BorrowMarket/index.spec.tsx
@@ -3,7 +3,7 @@ import BigNumber from 'bignumber.js';
 import { fireEvent } from '@testing-library/react';
 import { assetData } from '__mocks__/models/asset';
 import renderComponent from 'testUtils/renderComponent';
-import { useUserMarketInfo } from 'clients/api';
+import { useGetUserMarketInfo } from 'clients/api';
 import en from 'translation/translations/en.json';
 import BorrowMarket from '.';
 
@@ -11,10 +11,13 @@ jest.mock('clients/api');
 
 describe('pages/SupplyMarket', () => {
   beforeEach(() => {
-    (useUserMarketInfo as jest.Mock).mockImplementation(() => ({
-      assets: assetData,
-      userTotalBorrowLimitCents: new BigNumber('111'),
-      userTotalBorrowBalanceCents: new BigNumber('91'),
+    (useGetUserMarketInfo as jest.Mock).mockImplementation(() => ({
+      data: {
+        assets: assetData,
+        userTotalBorrowLimitCents: new BigNumber('111'),
+        userTotalBorrowBalanceCents: new BigNumber('91'),
+      },
+      isLoading: false,
     }));
   });
