+++ src/pages/Dashboard/Modals/BorrowRepay/Borrow/index.spec.tsx
@@ -8,7 +8,7 @@ import { SAFE_BORROW_LIMIT_PERCENTAGE } from 'config';
 import { Asset } from 'types';
 import fakeTransactionReceipt from '__mocks__/models/transactionReceipt';
 import fakeAccountAddress from '__mocks__/models/address';
-import { useUserMarketInfo, borrowVToken } from 'clients/api';
+import { useGetUserMarketInfo, borrowVToken } from 'clients/api';
 import useSuccessfulTransactionModal from 'hooks/useSuccessfulTransactionModal';
 import renderComponent from 'testUtils/renderComponent';
 import en from 'translation/translations/en.json';
@@ -29,10 +29,13 @@ jest.mock('hooks/useSuccessfulTransactionModal');
 
 describe('pages/Dashboard/BorrowRepayModal/Borrow', () => {
   beforeEach(() => {
-    (useUserMarketInfo as jest.Mock).mockImplementation(() => ({
-      assets: [], // Not used in these tests
-      userTotalBorrowLimitCents: fakeUserTotalBorrowLimitCents,
-      userTotalBorrowBalanceCents: fakeUserTotalBorrowBalanceCents,
+    (useGetUserMarketInfo as jest.Mock).mockImplementation(() => ({
+      data: {
+        assets: [], // Not used in these tests
+        userTotalBorrowLimitCents: fakeUserTotalBorrowLimitCents,
+        userTotalBorrowBalanceCents: fakeUserTotalBorrowBalanceCents,
+      },
+      isLoading: false,
     }));
   });
