+++ src/pages/Market/MarketTable/index.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import BigNumber from 'bignumber.js';
 import renderComponent from 'testUtils/renderComponent';
-import { useUserMarketInfo } from 'clients/api';
+import { useGetUserMarketInfo } from 'clients/api';
 import { assetData } from '__mocks__/models/asset';
 import MarketTable from '.';
 
@@ -9,11 +9,14 @@ jest.mock('clients/api');
 
 describe('pages/Market/MarketTable', () => {
   beforeEach(() => {
-    (useUserMarketInfo as jest.Mock).mockImplementation(() => ({
-      assets: assetData,
-      userTotalBorrowLimitCents: new BigNumber('111'),
-      userTotalBorrowBalanceCents: new BigNumber('91'),
-      userTotalSupplyBalanceCents: new BigNumber('910'),
+    (useGetUserMarketInfo as jest.Mock).mockImplementation(() => ({
+      data: {
+        assets: assetData,
+        userTotalBorrowLimitCents: new BigNumber('111'),
+        userTotalBorrowBalanceCents: new BigNumber('91'),
+        userTotalSupplyBalanceCents: new BigNumber('910'),
+      },
+      isLoading: false,
     }));
   });
