+++ src/pages/Xvs/index.spec.tsx
@@ -1,20 +1,23 @@
 import React from 'react';
 import BigNumber from 'bignumber.js';
 import renderComponent from 'testUtils/renderComponent';
-import { useUserMarketInfo } from 'clients/api';
+import { useGetUserMarketInfo } from 'clients/api';
 import Xvs from '.';
 
 jest.mock('clients/api');
 
 describe('pages/Xvs', () => {
   beforeEach(() => {
-    (useUserMarketInfo as jest.Mock).mockImplementation(() => ({
-      assets: [],
-      userTotalBorrowLimitCents: new BigNumber('111'),
-      userTotalBorrowBalanceCents: new BigNumber('91'),
-      userTotalSupplyBalanceCents: new BigNumber('910'),
-      dailyVenus: new BigNumber('22222'),
-      totalXvsDistributedWei: new BigNumber('22222'),
+    (useGetUserMarketInfo as jest.Mock).mockImplementation(() => ({
+      data: {
+        assets: [],
+        userTotalBorrowLimitCents: new BigNumber('111'),
+        userTotalBorrowBalanceCents: new BigNumber('91'),
+        userTotalSupplyBalanceCents: new BigNumber('910'),
+        dailyVenus: new BigNumber('22222'),
+        totalXvsDistributedWei: new BigNumber('22222'),
+      },
+      isLoading: false,
     }));
   });
